requirements.txt
# requirements.txt
Flask==2.3.3
Flask-PyMongo==2.3.0
PyMongo[srv]==4.6.1
Flask-JWT-Extended==4.5.3
python-dotenv==1.0.0
Werkzeug==2.3.8
requests==2.31.0
beautifulsoup4==4.12.3
lxml==5.1.0 # Parser for BeautifulSoup